add option_display_group.py
