+++ nixui/graphics/nav_interface.py
@@ -1,10 +1,10 @@
 import collections
 
-from PyQt5 import QtWidgets, QtCore
+from PyQt5 import QtWidgets
 
 from nixui.options import api, types
 from nixui.options.attribute import Attribute
-from nixui.graphics import generic_widgets, navbar, navlist, option_display
+from nixui.graphics import generic_widgets, navbar, navlist, option_display_group
 from nixui.utils.logger import logger
 
 
@@ -129,7 +129,7 @@ class OptionNavigationInterface(QtWidgets.QWidget):
                 )
             )
             self.fields_view.replace_widget(
-                FieldsGroupBox(
+                option_display_group.OptionDisplayGroupBox(
                     self.statemodel,
                     self.set_option_path,
                     option_path,
@@ -151,61 +151,3 @@ class OptionNavigationInterface(QtWidgets.QWidget):
             )
         )
         self.fields_view.replace_widget(QtWidgets.QLabel(''))
-
-
-class FieldsGroupBox(QtWidgets.QWidget):
-    def __init__(self, statemodel, set_option_path_fn, option=None, is_base_viewer=True, only_display_parent=False):
-        super().__init__()
-
-        self.elements = []
-
-        group_box = QtWidgets.QGroupBox()
-
-        if only_display_parent:
-            option_paths = [option]
-        else:
-            option_paths = api.get_option_tree().children(option)
-
-        if option_paths:
-            group_box.setTitle(str(option))
-        else:
-            option_paths = [option]
-
-        vbox = QtWidgets.QVBoxLayout()
-        for child_option_path in option_paths:
-            if not only_display_parent and len(api.get_option_tree().children(child_option_path)) > 1:
-                fields_group_box = FieldsGroupBox(
-                    statemodel,
-                    set_option_path_fn,
-                    child_option_path,
-                    is_base_viewer=False
-                )
-                self.elements.append(fields_group_box)
-                vbox.addWidget(fields_group_box)
-            else:
-                option_disp = option_display.GenericOptionDisplay(
-                    statemodel,
-                    set_option_path_fn,
-                    child_option_path
-                )
-                self.elements.append(option_disp)
-                vbox.addWidget(option_disp)
-            vbox.addWidget(generic_widgets.SeparatorLine())
-
-        if is_base_viewer:
-            vbox.addStretch()
-        group_box.setLayout(vbox)
-
-        lay = QtWidgets.QHBoxLayout()
-        if is_base_viewer:
-            scroll_area = QtWidgets.QScrollArea()
-            scroll_area.setWidgetResizable(True)
-            scroll_area.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOff)
-            lay.addWidget(scroll_area)
-            scroll_area.setWidget(group_box)
-        else:
-            lay.addWidget(group_box)
-
-        vbox.setContentsMargins(0, 0, 0, 0)
-        vbox.setSpacing(0)
-        self.setLayout(lay)
